Addiny isync
offlineimap proved to be too problematic in this set up for a bunch of reasons I
don't want to dwell upon. It's also worth mentioning that isync doesn't support
IMAP IDLE, so I have had to put in a 15 minute cron job to fire it off. No idea
how this will work in the scenario where there is limited or no internet
connectivity.

Feels like I'm banging a bunch of rocks together here sometimes in open source.